change: moving fixed point function to another file
